fix(guildDelete)

+++ src/events/client/guildDelete.js
@@ -9,7 +9,7 @@ module.exports = {
     once: false,
     async execute(client, guild) {
         //Bot has been kicked or banned in a guild
-        await client.db.deleteGuild(guild.id);
+        if (guild.available) await client.db.deleteGuild(guild.id);
         const bots = guild.available
             ? guild.members.cache.filter((m) => m.user.bot).size
             : null;
